fix compile errors
