+++ crates/proto/src/quic/mod.rs
@@ -14,10 +14,11 @@ mod quic_stream;
 
 pub use self::quic_client_stream::{
     client_config_tls13_webpki_roots, QuicClientConnect, QuicClientResponse, QuicClientStream,
-    QuicClientStreamBuilder, QuicLocalAddr,
+    QuicClientStreamBuilder,
 };
 pub use self::quic_server::{QuicServer, QuicStreams};
 pub use self::quic_stream::{DoqErrorCode, QuicStream};
+pub use crate::udp::QuicLocalAddr;
 
 #[cfg(test)]
 mod tests;
